internal-enrichment/censys-enrichment/tests/test-requirements.txt
# Main dependencies needs to be installed
-r ../requirements.txt
pytest==8.4.2
pytest-mock==3.15.1
factory-boy==3.3.3
